Read multiple files into a list with your favorite reader function.

Read multiple files into a list with your favorite reader function.

list_any(path = ".", .f, regexp = NULL, ignore.case = FALSE,
  invert = FALSE, ...)

Arguments

path
A character vector of one or more paths.

.f
A function able to read the desired file format.

regexp
A regular expression (e.g. [.]csv$) passed on to grep() to filter paths.

ignore.case
if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.

invert
If TRUE return files which do not match

...
Additional arguments passed to .f.

Value

A list.

See also

Examples

tor_example()

#> [1] "csv"   "mixed" "rdata" "rds"   "tsv"  

path <- tor_example("csv")
dir(path)

#> [1] "csv1.csv" "csv2.csv"

list_any(path, read.csv)

#> $csv1
#> # A tibble: 2 x 1
#>       x
#>   <int>
#> 1     1
#> 2     2
#> 
#> $csv2
#> # A tibble: 2 x 1
#>   y    
#>   <fct>
#> 1 a    
#> 2 b    
#> 

list_any(path, ~ read.csv(.x, stringsAsFactors = FALSE))

#> $csv1
#> # A tibble: 2 x 1
#>       x
#>   <int>
#> 1     1
#> 2     2
#> 
#> $csv2
#> # A tibble: 2 x 1
#>   y    
#>   <chr>
#> 1 a    
#> 2 b    
#> 

(path_mixed <- tor_example("mixed"))

#> [1] "C:/Users/LeporeM/Documents/Dropbox/git/ml/tor/inst/extdata/mixed"

dir(path_mixed)

#> [1] "csv.csv"           "lower_rdata.rdata" "rda.rda"          
#> [4] "upper_rdata.RData"

list_any(
  path_mixed, ~ get(load(.x)),
  regexp = "[.]csv$",
  invert = TRUE
)

#> $lower_rdata
#> # A tibble: 2 x 1
#>   y    
#>   <chr>
#> 1 a    
#> 2 b    
#> 
#> $rda
#> # A tibble: 2 x 1
#>   y    
#>   <chr>
#> 1 a    
#> 2 b    
#> 
#> $upper_rdata
#> # A tibble: 2 x 1
#>   y    
#>   <chr>
#> 1 a    
#> 2 b    
#> 

list_any(
  path_mixed, ~ get(load(.x)),
  "[.]Rdata$",
  ignore.case = TRUE
)

#> $lower_rdata
#> # A tibble: 2 x 1
#>   y    
#>   <chr>
#> 1 a    
#> 2 b    
#> 
#> $upper_rdata
#> # A tibble: 2 x 1
#>   y    
#>   <chr>
#> 1 a    
#> 2 b    
#>